auto_update/requirements.txt
feedparser
requests
beautifulsoup4
jinja2
openai
python-dateutil
schedule
deep-translator
